Rename existed "copilot" to "agent-copilot" (#1821)
