+++ spx-gui/src/App.vue
@@ -2,9 +2,9 @@
   <UIConfigProvider :config="config">
     <UIMessageProvider>
       <UIModalProvider>
-        <CopilotProvider>
+        <AgentCopilotProvider>
           <RouterView />
-        </CopilotProvider>
+        </AgentCopilotProvider>
       </UIModalProvider>
     </UIMessageProvider>
   </UIConfigProvider>
@@ -13,7 +13,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { UIConfigProvider, UIModalProvider, UIMessageProvider, type Config } from '@/components/ui'
-import CopilotProvider from '@/components/copilot/CopilotProvider.vue'
+import AgentCopilotProvider from '@/components/agent-copilot/CopilotProvider.vue'
 import { useI18n } from './utils/i18n'
 
 const { t } = useI18n()
