+++ spx-gui/src/components/agent-copilot/CopilotInput.vue
