+++ spx-gui/src/components/agent-copilot/CopilotUI.vue
@@ -2,8 +2,8 @@
 import { computed, nextTick, ref, watch } from 'vue'
 import { useBottomSticky } from '@/utils/dom'
 import { UIIcon } from '@/components/ui'
-import CopilotInput from '@/components/copilot/CopilotInput.vue'
-import CopilotRound from '@/components/copilot/CopilotRound.vue'
+import CopilotInput from './CopilotInput.vue'
+import CopilotRound from './CopilotRound.vue'
 import logoSrc from '../editor/code-editor/ui/copilot/logo.png'
 import envSvg from './env.svg?raw'
 import type { CopilotController } from './index'
