+++ spx-gui/src/components/agent-copilot/markdown/CodeView.vue
