+++ spx-gui/src/components/agent-copilot/markdown/FileContent.vue
