+++ spx-gui/src/components/agent-copilot/markdown/FileDiagnostics.vue
