+++ spx-gui/src/components/agent-copilot/markdown/MarkdownView.vue
