+++ spx-gui/src/components/agent-copilot/mcp/client.ts
