+++ spx-gui/src/components/agent-copilot/mcp/collector.ts
