+++ spx-gui/src/components/agent-copilot/mcp/definitions.ts
