+++ spx-gui/src/components/agent-copilot/mcp/registry.ts
