+++ spx-gui/src/components/agent-copilot/mcp/server.ts
