+++ spx-gui/src/components/agent-copilot/mcp/transport.ts
